Backend/requirements.txt
Django==5.2.5
djangorestframework==3.15.2
django-cors-headers==4.3.1
mongoengine==0.28.2
python-dotenv==1.0.0
djangorestframework-simplejwt==5.3.0
gunicorn==21.2.0
whitenoise==6.6.0
